backend/requirements.txt
# Core
pymupdf>=1.23.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pillow>=10.0.0

# AI & ML
openai>=1.0.0
tiktoken>=0.5.0
numpy>=1.24.0

# Database
pymongo>=4.6.0
